+++ WordService/pkg/service/word_service.go
@@ -90,7 +90,7 @@ func (wordService wordService) parseHtml(
 	wordMeangins := []model.WordMeaning{}
 
 	// 排序用的編號
-	var orderByNo int32 = 0
+	var orderByNo int64 = 0
 
 	c := colly.NewCollector(
 		colly.AllowedDomains("www.ldoceonline.com"),
